Bug/fix index (#978)
* fix index issues
