+++ src/components/Indexer/crawlerThread.ts
@@ -187,6 +187,9 @@ export async function processNetworkData(
             startBlock,
             blocksToProcess
           )
+          INDEXER_LOGGER.debug(
+            `Processed ${processedBlocks.foundEvents.length} events from ${chunkEvents.length} logs`
+          )
           currentBlock = await updateLastIndexedBlockNumber(
             processedBlocks.lastBlock,
             lastIndexedBlock
@@ -198,26 +201,25 @@ export async function processNetworkData(
           checkNewlyIndexedAssets(processedBlocks.foundEvents)
           chunkSize = chunkSize !== 1 ? chunkSize : rpcDetails.chunkSize
         } catch (error) {
-          INDEXER_LOGGER.log(
-            LOG_LEVELS_STR.LEVEL_ERROR,
-            `Processing event from network failed network: ${rpcDetails.network} Error: ${error.message} `,
-            true
+          INDEXER_LOGGER.error(
+            `Processing event from network failed network: ${rpcDetails.network} Error: ${error.message} `
           )
-          await updateLastIndexedBlockNumber(
-            startBlock + blocksToProcess,
-            lastIndexedBlock
-          )
+          // since something went wrong, we will not update the last indexed block
+          // so we will try to process the same chunk again
+          // after some sleep
+          await sleep(interval)
         }
+      } else {
+        await sleep(interval)
       }
       await processReindex(provider, signer, rpcDetails.chainId)
       lockProccessing = false
     } else {
       INDEXER_LOGGER.logMessage(
-        `Processing already in progress for network ${rpcDetails.network}, waiting until finishing the current processing ...`,
-        true
+        `Processing already in progress for network ${rpcDetails.network}, waiting until finishing the current processing ...`
       )
     }
-    await sleep(interval)
+
     // reindex chain command called
     if (REINDEX_BLOCK && !lockProccessing) {
       const networkHeight = await getNetworkHeight(provider)
